+++ modules/body-common/src/txrpc/remote/common/body/HttpId.java
@@ -5,13 +5,13 @@ import java.io.Serializable;
 public final class HttpId implements Serializable {
 
     public final String sessionId;
-    public final Long transactionId;
+    public final Integer transactionId;
 
     public HttpId() {
         this(null, null);
     }
 
-    public HttpId(String sessionId, Long transactionId) {
+    public HttpId(String sessionId, Integer transactionId) {
         this.sessionId = sessionId;
         this.transactionId = transactionId;
     }
