+++ modules/body-server/src/txrpc/remote/server/body/ServerHttpId.java
@@ -5,9 +5,9 @@ import txrpc.remote.server.IServerSessionId;
 public final class ServerHttpId {
 
     public final IServerSessionId sessionId;
-    public final Long transactionId;
+    public final Integer transactionId;
 
-    public ServerHttpId(IServerSessionId sessionId, Long transactionId) {
+    public ServerHttpId(IServerSessionId sessionId, Integer transactionId) {
         this.sessionId = sessionId;
         this.transactionId = transactionId;
     }
